refactor(tests): move call_tool_raw_seq to common harness, deduplicate circuit breaker logic
